requirements.txt
Flask==3.1.0
flask-cors==5.0.1
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.3
python-dotenv==1.1.0
supabase==2.13.0
joblib==1.4.2
